Fixing exception thrown by the client code under certain circumstances.
Occasionally the API will respond with HTML content instead of valid JSON content. This was causing an exception within the client code. This patch catches the exception and retries the request if the last attempt failed.

Major caveats of this patch include the fact that it retries the request instantly instead of waiting. Additionally there is no cutoff, so in cases where the PandaStream API has gone offline or is broken it will just sit there querying the PandaStream API repeatedly instead of using a proper backoff technique followed by eventually reporting a service error back to the use of the client.